requirements.txt
fastapi
uvicorn
anthropic
pypdf
python-multipart
httpx
fuzzywuzzy
python-Levenshtein
python-dotenv
pydantic
streamlit
crewai
langchain-google-genai
requests
python-whois
googlemaps
google-search-results